+++ openmetadata-ui/src/main/resources/ui/src/utils/DataInsightUtils.tsx
@@ -11,213 +11,17 @@
  *  limitations under the License.
  */
 
-import {
-  first,
-  get,
-  isInteger,
-  isString,
-  isUndefined,
-  last,
-  meanBy,
-  round,
-  sumBy,
-  toNumber,
-} from 'lodash';
-import { DateTime } from 'luxon';
-import { RangePickerProps } from '../components/common/DatePicker/DatePicker';
-import { PLACEHOLDER_ROUTE_TAB, ROUTES } from '../constants/constants';
+import { first, get, isUndefined, meanBy, round, sumBy } from 'lodash';
 import {
   ENTITIES_SUMMARY_LIST,
   WEB_SUMMARY_LIST,
 } from '../constants/DataInsight.constants';
-import { SystemChartType } from '../enums/DataInsight.enum';
-import {
-  DataInsightChartResult,
-  DataInsightChartType,
-} from '../generated/dataInsight/dataInsightChartResult';
-import { DailyActiveUsers } from '../generated/dataInsight/type/dailyActiveUsers';
-import {
-  ChartValue,
-  DataInsightTabs,
-} from '../interface/data-insight.interface';
+import type { SystemChartType } from '../enums/DataInsight.enum';
+import type { DataInsightChartResult } from '../generated/dataInsight/dataInsightChartResult';
+import { DataInsightChartType } from '../generated/dataInsight/dataInsightChartResult';
 import { DataInsightCustomChartResult } from '../rest/DataInsightAPI';
-import { customFormatDateTime } from './date-time/DateTimeUtils';
 import { t, translateWithNestedKeys } from './i18next/LocalUtil';
 import { pluralize } from './StringUtils';
-
-export const getEntryFormattedValue = (
-  value: number | string | undefined,
-  isPercentage?: boolean
-) => {
-  let suffix = '';
-  if (isPercentage) {
-    suffix = '%';
-  }
-
-  if (!isUndefined(value)) {
-    if (isString(value)) {
-      return `${value}${suffix}`;
-    } else if (isInteger(value)) {
-      return `${value}${suffix}`;
-    } else {
-      return `${round(value, 2)}${suffix}`;
-    }
-  } else {
-    return '';
-  }
-};
-
-/**
- * takes timestamps and raw data as inputs and return the graph data by mapping timestamp
- * @param timestamps timestamps array
- * @param rawData graph rwa data
- * @returns graph data
- */
-const prepareGraphData = (
-  timestamps: string[],
-  rawData: (
-    | {
-        [x: string]: ChartValue;
-        timestamp: string;
-      }
-    | undefined
-  )[]
-) => {
-  return (
-    timestamps.map((timestamp) => {
-      return rawData.reduce((previous, current) => {
-        if (current?.timestamp === timestamp) {
-          return { ...previous, ...current };
-        }
-
-        return previous;
-      }, {});
-    }) || []
-  );
-};
-
-/**
- *
- * @param latestData latest chart data
- * @returns latest sum count for chart
- */
-const getLatestCount = (latestData = {}) => {
-  let total = 0;
-  const latestEntries = Object.entries(latestData ?? {});
-
-  for (const entry of latestEntries) {
-    // if key is 'timestamp' or 'timestampValue' skipping its count for total
-    if (!['timestamp', 'timestampValue'].includes(entry[0])) {
-      total += toNumber(entry[1]);
-    }
-  }
-
-  return total;
-};
-
-/**
- *
- * @param rawData raw chart data
- * @param dataInsightChartType chart type
- * @returns formatted chart for graph
- */
-const getGraphFilteredData = (
-  rawData: DataInsightChartResult['data'] = [],
-  dataInsightChartType: DataInsightChartType
-) => {
-  const entities: string[] = [];
-  const timestamps: string[] = [];
-
-  const filteredData = rawData
-    .map((data) => {
-      if (data.timestamp && data.entityType) {
-        let value;
-        const timestamp = customFormatDateTime(data.timestamp, 'MMM dd');
-        if (!entities.includes(data.entityType ?? '')) {
-          entities.push(data.entityType ?? '');
-        }
-
-        if (!timestamps.includes(timestamp)) {
-          timestamps.push(timestamp);
-        }
-
-        switch (dataInsightChartType) {
-          case DataInsightChartType.PageViewsByEntities:
-            value = data.pageViews;
-
-            break;
-
-          default:
-            break;
-        }
-
-        return {
-          timestamp: timestamp,
-          timestampValue: data.timestamp,
-          [data.entityType ?? '']: value,
-        };
-      }
-
-      return;
-    })
-    .filter(Boolean);
-
-  return { filteredData, entities, timestamps };
-};
-
-/**
- *
- * @param rawData raw chart data
- * @param dataInsightChartType chart type
- * @returns required graph data by entity type
- */
-export const getGraphDataByEntityType = (
-  rawData: DataInsightChartResult['data'] = [],
-  dataInsightChartType: DataInsightChartType
-) => {
-  const { filteredData, entities, timestamps } = getGraphFilteredData(
-    rawData,
-    dataInsightChartType
-  );
-
-  const graphData = prepareGraphData(timestamps, filteredData);
-  const latestData = last(graphData) as Record<string, number>;
-  const oldData = first(graphData);
-  const latestPercentage = toNumber(getLatestCount(latestData));
-  const oldestPercentage = toNumber(getLatestCount(oldData));
-
-  const relativePercentage = latestPercentage - oldestPercentage;
-
-  return {
-    data: graphData,
-    entities,
-    total: getLatestCount(latestData),
-    relativePercentage: (relativePercentage / oldestPercentage) * 100,
-    latestData,
-  };
-};
-
-export const getFormattedActiveUsersData = (
-  activeUsers: DailyActiveUsers[]
-) => {
-  const formattedData = activeUsers.map((user) => ({
-    ...user,
-    timestampValue: user.timestamp,
-    timestamp: customFormatDateTime(user.timestamp, 'MMM dd'),
-  }));
-
-  const latestCount = Number(last(formattedData)?.activeUsers);
-  const oldestCount = Number(first(formattedData)?.activeUsers);
-
-  const relativePercentage = ((latestCount - oldestCount) / oldestCount) * 100;
-
-  return {
-    data: formattedData,
-    total: latestCount,
-    relativePercentage,
-  };
-};
-
 export const getEntitiesChartSummary = (
   chartResults?: Record<SystemChartType, DataInsightCustomChartResult>
 ) => {
@@ -283,13 +87,6 @@ export const getWebChartSummary = (
   return updatedSummary;
 };
 
-export const getDisabledDates: RangePickerProps['disabledDate'] = (current) => {
-  // Can not select days before today
-  const today = DateTime.now().startOf('day');
-
-  return current ? current.startOf('day') < today : false;
-};
-
 export const getKpiResultFeedback = (day: number, isTargetMet: boolean) => {
   if (day > 0 && isTargetMet) {
     return t('message.kpi-target-achieved-before-time');
@@ -304,91 +101,16 @@ export const getKpiResultFeedback = (day: number, isTargetMet: boolean) => {
   }
 };
 
-export const getDataInsightPathWithFqn = (tab = DataInsightTabs.DATA_ASSETS) =>
-  ROUTES.DATA_INSIGHT_WITH_TAB.replace(PLACEHOLDER_ROUTE_TAB, tab);
-
-export const getOptionalDataInsightTabFlag = (tab: DataInsightTabs) => {
-  return {
-    showDataInsightSummary:
-      tab === DataInsightTabs.APP_ANALYTICS ||
-      tab === DataInsightTabs.DATA_ASSETS,
-    showKpiChart:
-      tab === DataInsightTabs.KPIS || tab === DataInsightTabs.DATA_ASSETS,
-  };
-};
-
-export const sortEntityByValue = (
-  entities: string[],
-  latestData: Record<string, number>
-) => {
-  const entityValues = entities.map((entity) => ({
-    entity,
-    value: latestData[entity] ?? 0,
-  }));
-
-  // Sort the entities based on their values in descending order
-  entityValues.sort((a, b) => b.value - a.value);
-
-  // Extract the sorted entities without their values
-  return entityValues.map((entity) => entity.entity);
-};
-
-export const getRandomHexColor = () => {
-  const randomColor = Math.floor(Math.random() * 16777215).toString(16);
-
-  return `#${randomColor}`;
-};
-
-export const isPercentageSystemGraph = (graph: SystemChartType) => {
-  return [
-    SystemChartType.PercentageOfDataAssetWithDescription,
-    SystemChartType.PercentageOfDataAssetWithOwner,
-    SystemChartType.PercentageOfServiceWithDescription,
-    SystemChartType.PercentageOfServiceWithOwner,
-  ].includes(graph);
-};
-
-export const getQueryFilterForDataInsightChart = (
-  teamFilter?: string[],
-  tierFilter?: string[]
-) => {
-  if (!tierFilter && !teamFilter) {
-    return undefined;
-  }
-
-  return JSON.stringify({
-    query: {
-      bool: {
-        must: [
-          ...(tierFilter
-            ? [
-                {
-                  bool: {
-                    should: tierFilter.map((tier) => ({
-                      term: { 'tier.keyword': tier },
-                    })),
-                  },
-                },
-              ]
-            : []),
-          ...(teamFilter
-            ? [
-                {
-                  nested: {
-                    path: 'owners',
-                    query: {
-                      bool: {
-                        should: teamFilter.map((team) => ({
-                          term: { 'owners.name.keyword': team },
-                        })),
-                      },
-                    },
-                  },
-                },
-              ]
-            : []),
-        ],
-      },
-    },
-  });
-};
+// Re-exports from DataInsightPureUtils (backward compat)
+export {
+  getDataInsightPathWithFqn,
+  getDisabledDates,
+  getEntryFormattedValue,
+  getFormattedActiveUsersData,
+  getGraphDataByEntityType,
+  getOptionalDataInsightTabFlag,
+  getQueryFilterForDataInsightChart,
+  getRandomHexColor,
+  isPercentageSystemGraph,
+  sortEntityByValue,
+} from './DataInsightPureUtils';
